refactor(etl): noyau pur de linéarisation etl/parsing/ + adapters DLT minces (#121)
Phase 1 du filet C3 (prérequis de la discovery #55) : la linéarisation
sélective des flux devient un noyau pur sans dépendance DLT, testable au
seam « bytes fichier → records » — les dicts émis sont, aux métadonnées
DLT près, les lignes des tables DuckDB.

- Nouveau package etl/parsing/ (0 import dlt) :
  - xml.py : ConfigFluxXml (@dataclass frozen slots, ADR-0018) chargée
    depuis flux.yaml avec allowlist de clés — les typos YAML explosent
    au chargement ; parser_flux_xml(bytes, config, traçabilité).
  - r64.py : parser_flux_r64(bytes, traçabilité), format wide.
  - tracabilite.py : TracabiliteFlux ; l'enrichissement (_source_zip,
    _flux_type, _xml/_json_name, modification_date) reste dans le noyau
    — il fait partie du contrat des tables.
- transformers/parsers.py : 721 → ~90 lignes, 2 factories @dlt.transformer
  minces qui déballent l'item DLT et portent la politique d'erreur
  d'ingestion (log + skip, re-tenté au prochain run).
- Suppressions (deletion test sur le corps, aucun clone vivant) :
  - chemin JSON générique : mort ET cassé (json_to_dict_from_bytes
    inexistant, noqa F821 ; aucune config ne pouvait l'atteindre) ;
  - chemin CSV : dormant, aucun csv_configs dans flux.yaml, aucun flux
    CSV anticipé ; branche sftp_enedis + diagnostic_flux nettoyées,
    transformer_type ≠ r64_timeseries lève désormais explicitement ;
  - constantes unzip_xml/csv_transformer pré-configurées sans appelant.
- Vérifié : les 6 Configurations de flux réelles passent l'allowlist,
  smoke du parser pur avec la config C15 réelle, import source OK.
- etl/CONTEXT.md : entrées Linéarisation sélective, Configuration de
  flux ; Transformer redéfini comme adapter mince autour du noyau pur.

Phase 2 (#121) : batterie de caractérisation golden dès que les fichiers
réels anonymisés sont fournis (HITL).
