+++ electricore/etl/CONTEXT.md
@@ -43,4 +43,11 @@ Exécution asynchrone du pipeline d'ingestion déclenchée via l'API (`POST /etl
 Conteneur Docker qui déclenche périodiquement l'ETL via cron et un appel HTTP à `/etl/run` — pas de logique métier embarquée. Voir [ADR-0011](../../docs/adr/0011-deploiement-vps-docker.md).
 
 **Transformer** :
-Étape unitaire et composable de transformation d'un fichier (`crypto.py` déchiffrement AES, `archive.py` extraction ZIP, `parsers.py` parsing XML/CSV). Les transformers se chaînent via le `|` de DLT : `encrypted | decrypt | unzip | parse`.
+Étape unitaire et composable de transformation d'un fichier, sous forme d'*adapter* DLT mince autour d'un noyau pur (`crypto.py` déchiffrement AES, `archive.py` extraction ZIP, `parsers.py` linéarisation). Les transformers se chaînent via le `|` de DLT : `encrypted | decrypt | unzip | parse`.
+
+**Linéarisation sélective** :
+Transformation d'un document hiérarchique (XML, JSON) en enregistrements plats — une ligne par occurrence du `row_level` — en n'extrayant *que* les champs décrits par la *Configuration de flux*. Les enregistrements produits sont, aux métadonnées DLT près, les lignes des tables DuckDB. Implémentée par le noyau pur `etl/parsing/` (sans dépendance DLT), héritée de l'ancienne lib `electriflux` (cf. [ADR-0001](../../docs/adr/0001-monorepo.md)).
+_Éviter_ : extraction (collision avec le E de ETL = récupération SFTP), parsing tout court (la sélection est constitutive).
+
+**Configuration de flux** :
+Entrée de `flux.yaml` décrivant, pour un type de flux : le `file_pattern` SFTP, et par table cible le `row_level` (XPath des lignes), les mappings champ → XPath (`metadata_fields`, `data_fields`) et les `nested_fields` conditionnels (ex : `Code_Qualification 1/2` → préfixes `avant_`/`apres_`). C'est le **contrat de sélection** de la linéarisation — ce qui n'y figure pas n'entre pas dans DuckDB.
